refactor: move `nonNullable` to shared
