+++ packages/shared/src/array.ts
