+++ packages/unminify/src/index.ts
@@ -1,9 +1,9 @@
+import { nonNullable } from '@wakaru/shared/array'
 import { jscodeshiftWithParser as j, printSourceWithErrorLoc } from '@wakaru/shared/jscodeshift'
 import { Timing } from '@wakaru/shared/timing'
 import { basename } from 'pathe'
 import { transformationRules } from './transformations'
 import { arraify } from './utils/arraify'
-import { nonNullable } from './utils/utils'
 import type { MaybeArray } from './utils/arraify'
 import type { TransformationRule } from '@wakaru/shared/rule'
 import type { Collection, FileInfo } from 'jscodeshift'
