+++ packages/unminify/src/transformations/smart-inline.ts
@@ -3,8 +3,8 @@ import { generateName } from '@wakaru/ast-utils/identifier'
 import { createObjectProperty } from '@wakaru/ast-utils/object'
 import { findReferences } from '@wakaru/ast-utils/reference'
 import { MultiMap } from '@wakaru/ds'
+import { nonNullable } from '@wakaru/shared/array'
 import { createJSCodeshiftTransformationRule } from '@wakaru/shared/rule'
-import { nonNullable } from '../utils/utils'
 import type { ASTTransformation } from '@wakaru/shared/rule'
 import type { CommentKind, StatementKind } from 'ast-types/lib/gen/kinds'
 import type { Scope } from 'ast-types/lib/scope'
